+++ src/blas_like/level1/CopyETI.cpp.in
@@ -1,6 +1,5 @@
 #include "El/core.hpp"
 #include "El/blas_like/level1/Copy.hpp"
-#include "El/blas_like/level1/CopyAsync.hpp"
 
 namespace El
 {
